Make basic specification reusable
