+++ src/Spec/VocaLinkV380/DataInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Cs278\BankModulus\Spec\VocaLinkV380;
+
+/**
+ * @internal This class is not part of the public API of this package.
+ */
+interface DataInterface
+{
+    public function fetchRecord($sortCode, $pass);
+}
