+++ tests/Spec/VocaLinkV380Test.php
@@ -7,7 +7,8 @@ use Cs278\BankModulus\BankAccountNormalizer\DefaultNormalizer;
 
 /**
  * @covers Cs278\BankModulus\Spec\VocaLinkV380
- * @covers Cs278\BankModulus\Spec\VocaLinkV380Data
+ * @covers Cs278\BankModulus\Spec\VocaLinkV380\Driver
+ * @covers Cs278\BankModulus\Spec\VocaLinkV380\DataV380
  */
 final class VocaLinkV380Test extends \PHPUnit_Framework_TestCase
 {
